feat: adicionando configuração para uso de DTO
