+++ AgendamentoBanhoETosa/Model/Entities/Servico.cs
@@ -1,0 +1,10 @@
+namespace AgendamentoBanhoETosa.Model
+{
+    public class Servico
+    {
+        public int Id { get; set; }
+        public string Nome { get; set; }
+        public float Preco { get; set; }
+
+    }
+}
